Add descriptions to commands and improve REPL display formatting
