+++ engine/src/core/repl/commands.rs
@@ -1,34 +1,34 @@
+use super::COMMAND_LIST;
 use std::process::Command;
-
 use log2::{debug, info};
 
-use crate::core::repl::COMMAND_LIST;
-
-pub fn say_hello() {
-    println!("Hello from your new command!");
+pub(crate) fn say_hello() {
+    println!("Hello, World!");
 }
 
-pub fn echo(args: Vec<String>) {
+pub(crate) fn echo(args: Vec<String>) {
     debug!("{}", args.join(" "));
     println!("{}", args.join(" "))
 }
 
-pub fn exit() {
+pub(crate) fn exit() {
     debug!("Exiting...");
     std::process::exit(0)
 }
-pub fn clear() {
+
+pub(crate) fn clear() {
     info!("Clearing screen..., running command");
     let _result = if cfg!(target_os = "windows") {
         debug!("target_os is windows");
         Command::new("cmd").args(["/c", "cls"]).spawn()
     } else {
-        debug!("target_os was unix");
+        debug!("target_os is unix");
         // "clear" or "tput reset"
         Command::new("tput").arg("reset").spawn()
     };
 }
-pub fn cmds() {
+
+pub(crate) fn help() {
     println!("Commands:");
     for cmd in COMMAND_LIST.commands.read().iter() {
         println!("{:#}", cmd);
